requirements.txt
PyQt5
pandas
numpy
matplotlib
seaborn
python-docx
fpdf
PyPDF2
scikit-learn
spacy
